requirements.txt
python-telegram-bot==13.15
environs==11.0.0
google-cloud-dialogflow==2.38.0
vk-api==11.9.9
